refactor: project 도메인에서만 사용하여 파일 이동 및 네이밍 변경
